:art: 토큰 저장 장소 변경

+++ app/api/auth/logout/route.ts
@@ -3,15 +3,27 @@ import { validateAccessToken } from "@/lib/validateToken";
 import prisma from "@/lib/db";
 
 export async function DELETE(req: NextRequest) {
-  const accessToken = req.headers.get("Authorization")?.split(" ")[1];
+  const refreshToken = req.cookies.get("refreshToken");
 
-  const decoded = validateAccessToken(accessToken as string);
+  if (!refreshToken) {
+    return NextResponse.json(
+      { message: "로그인되지 않은 사용자입니다." },
+      { status: 401 },
+    );
+  }
 
-  const userId = decoded.uid;
+  const decoded = validateAccessToken(refreshToken.value);
 
   await prisma.refreshToken.delete({
-    where: { userId },
+    where: { token: refreshToken.value },
   });
 
-  return NextResponse.json({ message: "로그아웃되었습니다." });
+  return NextResponse.json(
+    { message: "로그아웃이 완료되었습니다." },
+    {
+      headers: {
+        "Set-Cookie": "refreshToken=; Path=/ ; HttpOnly; Max-Age=0",
+      },
+    },
+  );
 }
